# CascadeProjects/city-lifestyle-app/infrastructure/gcp/terraform/core/api_gateway/specs/openapi.yaml
swagger: '2.0'
info:
  title: City Lifestyle API
  description: API Gateway configuration for City Lifestyle application
  version: 1.0.0
schemes:
  - https
produces:
  - application/json
securityDefinitions:
  api_key:
    type: apiKey
    name: x-api-key
    in: header
  jwt:
    type: oauth2
    authorizationUrl: ''
    flow: implicit
    x-google-issuer: 'https://securetoken.google.com/${project_id}'
    x-google-jwks_uri: 'https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com'
    x-google-audiences: '${project_id}'
security:
  - api_key: []
    jwt: []
x-google-backend:
  address: 'https://${region}-run.googleapis.com'
  path_translation: APPEND_PATH_TO_ADDRESS
paths:
  /api/v1/users:
    get:
      summary: List users
      operationId: listUsers
      security:
        - jwt: []
      x-google-backend:
        address: 'https://${region}-run.googleapis.com/users'
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create user
      operationId: createUser
      security:
        - jwt: []
      x-google-backend:
        address: 'https://${region}-run.googleapis.com/users'
      responses:
        '201':
          description: User created
        '400':
          description: Invalid request
        '401':
          description: Unauthorized

  /api/v1/auth:
    post:
      summary: Authenticate user
      operationId: authenticate
      security:
        - api_key: []
      x-google-backend:
        address: 'https://${region}-run.googleapis.com/auth'
      responses:
        '200':
          description: Authentication successful
        '401':
          description: Authentication failed

  /api/v1/places:
    get:
      summary: List places
      operationId: listPlaces
      security:
        - jwt: []
      x-google-backend:
        address: 'https://${region}-run.googleapis.com/places'
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
    post:
      summary: Create place
      operationId: createPlace
      security:
        - jwt: []
      x-google-backend:
        address: 'https://${region}-run.googleapis.com/places'
      responses:
        '201':
          description: Place created
        '400':
          description: Invalid request
        '401':
          description: Unauthorized

  /api/v1/events:
    get:
      summary: List events
      operationId: listEvents
      security:
        - jwt: []
      x-google-backend:
        address: 'https://${region}-run.googleapis.com/events'
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
    post:
      summary: Create event
      operationId: createEvent
      security:
        - jwt: []
      x-google-backend:
        address: 'https://${region}-run.googleapis.com/events'
      responses:
        '201':
          description: Event created
        '400':
          description: Invalid request
        '401':
          description: Unauthorized

  /health:
    get:
      summary: Health check
      operationId: healthCheck
      x-google-backend:
        address: 'https://${region}-run.googleapis.com/health'
      responses:
        '200':
          description: Service is healthy